Support float type in mitmweb
This is a parital fix for #3144. Adds float to the to_str, and a new
Option in React. There is however an outstanding issue.

Javascript truncates whole numbers from parseFloat leading to Python
interpreting them as integers. E.g parseFloat(1.0) -> 1 (which Python
sees as an integer). May need to coerce int to floating point in order
to address this properly.